Projects/PawelPawlik_MichalSiennicki_AlicjaZiarko/checkpoint3/src/requirements.txt
torch==1.13.0
torchvision==0.14.0
torchaudio==0.13.0
hydra-core==1.2.0
matplotlib==3.6.2
pandas==1.5.1
kaggle==1.5.12
notebook==6.5.2
notebook_shim==0.2.2
opencv-python==4.5.5.64
black==22.10.0
torchmetrics==0.9.0
numpy==1.22.4
Pillow==9.1.1
hydra-core==1.2.0
scikit-learn==1.1.2
pytorch-lightning==1.8.3.post1
wandb==0.13.6
zennit==0.5.0
ipyplot==1.1.1
imageio==2.25.0
